+++ ornitho/model/modification_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class ModificationType(Enum):
+    ONLY_MODIFIED = "only_modified"
+    ONLY_DELETED = "only_deleted"
+    ALL = "all"